move function to start a node in test
